requirements.txt
pandas
jupyter
notebook
matplotlib
scikit-learn
xgboost
interpret
ydata-profiling
openpyxl
sweetviz
scipy
dvc
mlflow
dvc_s3
python-dotenv